Move syntax highlighting to shared module
