+++ frontend/src/metabase/query_builder/components/expressions/HighlightExpression/HighlightExpression.module.css
@@ -7,66 +7,3 @@
   line-height: 1.065rem;
   overflow: auto;
 }
-
-.keyword {
-  color: var(--mb-color-saturated-purple);
-}
-
-.name,
-.variable {
-  color: var(--mb-color-saturated-blue);
-}
-
-.function.variableName {
-  font-weight: bold;
-}
-
-.constant,
-.typeName {
-  color: var(--mb-color-saturated-red);
-}
-
-.string {
-  color: var(--mb-color-saturated-green);
-}
-
-.number,
-.bool {
-  color: var(--mb-color-saturated-red);
-}
-
-.paren {
-  color: inherit;
-}
-
-.comment,
-.lineComment,
-.blockComment {
-  color: var(--mb-color-text-light);
-  font-style: italic;
-}
-
-.squareBracket {
-  color: var(--mb-color-brand);
-}
-
-.callExpression {
-  font-weight: bold;
-}
-
-.processingInstruction {
-  color: var(--mb-color-brand);
-}
-
-.logicOperator {
-  color: var(--mb-color-saturated-purple);
-}
-
-.arithmeticOperator,
-.compareOperator {
-  color: var(--mb-color-text-dark);
-}
-
-.escape {
-  color: var(--mb-color-saturated-red);
-}
